added error handling, code improved

+++ CLI/parser.hpp
@@ -7,7 +7,7 @@ class Parser
     SyntaxAnalyzer syntaxAnalyzer;   // compostion
     LexicalAnalyzer lexicalAnalyzer; // compostion
     
-    size_t position = 0;
+    int position = 0;
     
 public:
     void parse(std::string input);
